# docs/swagger.yaml
basePath: /
definitions:
...
  ent.FLink:
    properties:
      avatar_url:
        description: 头像
        type: string
      cover_url:
        description: 网站封面
        type: string
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      description:
        description: 简介
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/ent.FLinkEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the FLinkQuery when eager-loading is set.
      email:
        description: 邮箱
        type: string
      enable_friend_circle:
        description: 是否开启朋友圈
        type: boolean
      friend_circle_rule_id:
        description: 朋友圈解析规则
        type: integer
      group_id:
        description: 友链组
        type: integer
      id:
        description: ID of the ent.
        type: integer
      name:
        description: 名称
        type: string
      snapshot_url:
        description: 网站截图
        type: string
      status:
        description: 状态
        type: integer
      updated_at:
        description: UpdatedAt holds the value of the "updated_at" field.
        type: string
      url:
        description: 链接
        type: string
    type: object
  ent.FLinkEdges:
    properties:
      group:
        allOf:
        - $ref: '#/definitions/ent.FLinkGroup'
        description: Group holds the value of the group edge.
    type: object
  ent.FLinkGroup:
    properties:
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/ent.FLinkGroupEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the FLinkGroupQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: integer
      name:
        description: 名称
        type: string
      updated_at:
        description: UpdatedAt holds the value of the "updated_at" field.
        type: string
    type: object
  ent.FLinkGroupEdges:
    properties:
      links:
        description: Links holds the value of the links edge.
        items:
          $ref: '#/definitions/ent.FLink'
        type: array
    type: object
...
  ent.StorageStrategy:
    properties:
      access_key:
        description: accessKey
        type: string
      base_path:
        description: local 基础路径
        type: string
      bucket:
        description: 存储桶名称
        type: string
      created_at:
        description: CreatedAt holds the value of the "created_at" field.
        type: string
      domain:
        description: 访问域名
        type: string
      endpoint:
        description: 端点
        type: string
      id:
        description: ID of the ent.
        type: integer
      master:
        description: 是否为默认策略
        type: boolean
      name:
        description: 策略名称
        type: string
      node_id:
        description: 节点 ID
        type: string
      region:
        description: region
        type: string
      secret_key:
        description: secret_key
        type: string
      type:
        allOf:
        - $ref: '#/definitions/storagestrategy.Type'
        description: 策略类型
      updated_at:
        description: UpdatedAt holds the value of the "updated_at" field.
        type: string
    type: object
...
  model.HttpError:
    properties:
      code:
        type: integer
      msg:
        type: string
    type: object
  model.HttpSuccess:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  model.StorageStrategyCreateReq:
    properties:
      access_key:
        type: string
      base_path:
        type: string
      bucket:
        type: string
      domain:
        type: string
      endpoint:
        type: string
      master:
        type: boolean
      name:
        type: string
      region:
        type: string
      secret_key:
        type: string
      type:
        enum:
        - local
        - s3
        type: string
    required:
    - domain
    - name
    - type
    type: object
  model.StorageStrategyListResp:
    properties:
      id:
        type: integer
      master:
        type: boolean
      name:
        type: string
      type:
        type: string
    type: object
  model.StorageStrategyUpdateReq:
    properties:
      access_key:
        type: string
      base_path:
        type: string
      bucket:
        type: string
      domain:
        type: string
      endpoint:
        type: string
      master:
        type: boolean
      name:
        type: string
      node_id:
        type: string
      region:
        type: string
      secret_key:
        type: string
      type:
        enum:
        - local
        - s3
        type: string
    required:
    - domain
    - type
    type: object
...
  storagestrategy.Type:
    enum:
    - local
    - local
    - s3
    type: string
    x-enum-varnames:
    - DefaultType
    - TypeLocal
    - TypeS3
host: localhost:13000
info:
  contact:
    email: fiber@swagger.io
    name: API Support
  description: This is a sample swagger for Fiber
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fiber Example API
  version: "1.0"
paths:
...
  /api/v1/flink-group/create:
    post:
      consumes:
      - application/json
      description: 创建FlinkGroup
      parameters:
      - description: FlinkGroup创建请求体
        in: body
        name: flink_group_create_req
        required: true
        schema:
          $ref: '#/definitions/ent.FLinkGroup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/ent.FLinkGroup'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 创建FlinkGroup
      tags:
      - flinkGroup
  /api/v1/flink-group/delete/{id}:
    delete:
      consumes:
      - application/json
      description: 删除FlinkGroup
      parameters:
      - description: FlinkGroup ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 删除FlinkGroup
      tags:
      - flinkGroup
  /api/v1/flink-group/list:
    get:
      consumes:
      - application/json
      description: 查询FlinkGroup
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/ent.FLinkGroup'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 查询FlinkGroup
      tags:
      - flinkGroup
  /api/v1/flink-group/update/{id}:
    put:
      consumes:
      - application/json
      description: 更新FlinkGroup
      parameters:
      - description: FlinkGroup ID
        in: path
        name: id
        required: true
        type: integer
      - description: FlinkGroup更新请求体
        in: body
        name: flink_group_update_req
        required: true
        schema:
          $ref: '#/definitions/ent.FLinkGroup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/ent.FLinkGroup'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 更新FlinkGroup
      tags:
      - flinkGroup
...
  /api/v1/storage-strategy/create:
    post:
      consumes:
      - application/json
      description: 创建一个新的存储策略
      parameters:
      - description: Storage Strategy Create Request
        in: body
        name: strategy
        required: true
        schema:
          $ref: '#/definitions/model.StorageStrategyCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/ent.StorageStrategy'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 创建存储策略
      tags:
      - storageStrategy
  /api/v1/storage-strategy/default/{id}:
    put:
      consumes:
      - application/json
      description: 设置指定ID的存储策略为默认策略
      parameters:
      - description: Storage Strategy ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 设置默认存储策略
      tags:
      - storageStrategy
  /api/v1/storage-strategy/delete/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定ID的存储策略
      parameters:
      - description: Storage Strategy ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 删除存储策略
      tags:
      - storageStrategy
  /api/v1/storage-strategy/list:
    get:
      consumes:
      - application/json
      description: 获取所有存储策略的列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/ent.StorageStrategy'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 获取存储策略列表
      tags:
      - storageStrategy
  /api/v1/storage-strategy/list-all:
    get:
      consumes:
      - application/json
      description: 获取所有存储策略的列表，包括默认策略
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.StorageStrategyListResp'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 获取所有存储策略列表
      tags:
      - storageStrategy
  /api/v1/storage-strategy/query/{id}:
    get:
      consumes:
      - application/json
      description: 查询指定ID的存储策略
      parameters:
      - description: Storage Strategy ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/ent.StorageStrategy'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 查询存储策略
      tags:
      - storageStrategy
  /api/v1/storage-strategy/update/{id}:
    put:
      consumes:
      - application/json
      description: 更新指定ID的存储策略
      parameters:
      - description: Storage Strategy ID
        in: path
        name: id
        required: true
        type: integer
      - description: Storage Strategy Update Request
        in: body
        name: strategy
        required: true
        schema:
          $ref: '#/definitions/model.StorageStrategyUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/ent.StorageStrategy'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 更新存储策略
      tags:
      - storageStrategy
...
swagger: "2.0"